+++ ArcdpsLogManager/Configuration/DpsReportUploadSettingsPage.cs
@@ -10,7 +10,8 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 	{
 		private readonly RadioButtonList domainList;
 		private readonly CheckBox uploadDetailedWvwCheckbox;
-		
+		private readonly CheckBox autoUploadCheckbox;
+
 		private bool EditingUserToken { get; set; }
 
 		public DpsReportUploadSettingsPage()
@@ -43,49 +44,37 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 			int domainIndex = domainList.DataStore.TakeWhile(element => element != currentDomain).Count();
 			domainList.SelectedIndex = domainIndex;
 
-			var domainDescriptionLabel = new Label
-			{
-				Wrap = WrapMode.Word,
-				Height = 50,
-				Text = ((DpsReportDomain) domainList.SelectedValue).Description
-			};
+			var domainDescriptionLabel = new Label { Wrap = WrapMode.Word, Height = 50, Text = ((DpsReportDomain) domainList.SelectedValue).Description };
 
 			domainList.SelectedValueChanged += (sender, args) =>
 			{
 				domainDescriptionLabel.Text = ((DpsReportDomain) domainList.SelectedValue).Description;
 			};
 
-			uploadDetailedWvwCheckbox = new CheckBox
-			{
-				Text = "Detailed WVW Reports (Beta)",
-				Checked = Settings.DpsReportUploadDetailedWvw
-			};
+			uploadDetailedWvwCheckbox = new CheckBox { Text = "Detailed WVW Reports (large files may fail)", Checked = Settings.DpsReportUploadDetailedWvw };
 
-			var userTokenTextBox = new TextBox
-			{
-				ReadOnly = true,
-				Text = "************",
-				Enabled = false
-			};
+			autoUploadCheckbox = new CheckBox { Text = "Automatically upload logs", Checked = Settings.DpsReportAutoUpload };
 
-			var showUserTokenButton = new Button {Text = "Show"};
+			var userTokenTextBox = new TextBox { ReadOnly = true, Text = "************", Enabled = false };
+
+			var showUserTokenButton = new Button { Text = "Show" };
 			showUserTokenButton.Click += (sender, args) =>
 			{
 				userTokenTextBox.Text = Settings.DpsReportUserToken;
 				userTokenTextBox.Enabled = true;
 			};
-			var changeUserTokenButton = new Button {Text = "Change"};
+			var changeUserTokenButton = new Button { Text = "Change" };
 			changeUserTokenButton.Click += (_, _) =>
 			{
 				if (EditingUserToken)
 				{
 					EditingUserToken = false;
-					
+
 					Settings.DpsReportUserToken = userTokenTextBox.Text;
 					userTokenTextBox.ReadOnly = true;
 					userTokenTextBox.Text = "************";
 					userTokenTextBox.Enabled = false;
-					
+
 					changeUserTokenButton.Text = "Change";
 					showUserTokenButton.Visible = true;
 				}
@@ -105,18 +94,30 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 			var layout = new DynamicLayout();
 			layout.BeginVertical(new Padding(10), new Size(5, 5));
 			{
+				layout.BeginGroup("Automatic uploads", new Padding(5), new Size(5, 5));
+				{
+					layout.AddRow(new Label
+					{
+						Text = "If enabled, newly discovered logs will be automatically queued for an upload if they are less than 1 day old.",
+						Wrap = WrapMode.Word,
+					});
+					layout.AddRow(autoUploadCheckbox);
+				}
+				layout.EndGroup();
+
+				layout.BeginGroup("Upload options", new Padding(5), new Size(5, 5));
+				{
+					layout.AddRow(uploadDetailedWvwCheckbox);
+				}
+				layout.EndGroup();
+
 				layout.BeginGroup("Upload domain", new Padding(5), new Size(5, 5));
 				{
 					layout.AddRow(domainList);
 					layout.AddRow(domainDescriptionLabel);
 				}
 				layout.EndGroup();
-				layout.BeginGroup("Upload Options", new Padding(5), new Size(5, 5));
-				{
-					layout.AddRow(uploadDetailedWvwCheckbox);
-				}
-				layout.EndGroup();
-				
+
 				layout.BeginGroup("User token", new Padding(5), new Size(5, 5));
 				{
 					layout.BeginVertical();
@@ -126,7 +127,6 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 							Text = "The user token used for dps.report uploads. Treat is as a password, " +
 							       "it can be used to see all previous uploads.",
 							Wrap = WrapMode.Word,
-							Height = 50
 						});
 					}
 					layout.EndVertical();
@@ -156,6 +156,11 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 			if (uploadDetailedWvwCheckbox.Checked.HasValue)
 			{
 				Settings.DpsReportUploadDetailedWvw = uploadDetailedWvwCheckbox.Checked.Value;
+			}
+
+			if (autoUploadCheckbox.Checked.HasValue)
+			{
+				Settings.DpsReportAutoUpload = autoUploadCheckbox.Checked.Value;
 			}
 		}
 	}