+++ ArcdpsLogManager/Configuration/Settings.cs
@@ -38,6 +38,7 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 			CheckForUpdatesChanged += (sender, args) => SaveToFile();
 			DpsReportUserTokenChanged += (sender, args) => SaveToFile();
 			DpsReportDomainChanged += (sender, args) => SaveToFile();
+			DpsReportAutoUploadChanged += (sender, args) => SaveToFile();
 			MinimumLogDurationSecondsChanged += (sender, args) => SaveToFile();
 			IgnoredUpdateVersionsChanged += (sender, args) => SaveToFile();
 			DpsReportUploadDetailedWvwChanged += (sender, args) => SaveToFile();
@@ -266,6 +267,19 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 				}
 			}
 		}
+		
+		public static bool DpsReportAutoUpload
+		{
+			get => Values.DpsReportAutoUpload;
+			set
+			{
+				if (Values.DpsReportAutoUpload != value)
+				{
+					Values.DpsReportAutoUpload = value;
+					OnDpsReportAutoUploadChanged();
+				}
+			}
+		}
 
 		public static bool DpsReportUploadDetailedWvw
 		{
@@ -316,6 +330,7 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 		public static event EventHandler<EventArgs> CheckForUpdatesChanged;
 		public static event EventHandler<EventArgs> DpsReportUserTokenChanged;
 		public static event EventHandler<EventArgs> DpsReportDomainChanged;
+		public static event EventHandler<EventArgs> DpsReportAutoUploadChanged;
 		public static event EventHandler<EventArgs> MinimumLogDurationSecondsChanged;
 		public static event EventHandler<EventArgs> IgnoredUpdateVersionsChanged;
 		public static event EventHandler<EventArgs> DpsReportUploadDetailedWvwChanged;
@@ -360,6 +375,11 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 		private static void OnDpsReportUserTokenChanged()
 		{
 			DpsReportUserTokenChanged?.Invoke(null, EventArgs.Empty);
+		}
+		
+		private static void OnDpsReportAutoUploadChanged()
+		{
+			DpsReportAutoUploadChanged?.Invoke(null, EventArgs.Empty);
 		}
 
 		private static void OnMinimumLogDurationSecondsChanged()