+++ ArcdpsLogManager/Processing/BackgroundProcessorEventArgs.cs
@@ -2,17 +2,11 @@ using System;
 
 namespace GW2Scratch.ArcdpsLogManager.Processing
 {
-	public class BackgroundProcessorEventArgs : EventArgs
+	public class BackgroundProcessorEventArgs<T>(T item, int currentScheduledItems, int totalProcessedItems, int totalScheduledItems) : EventArgs
 	{
-		public int TotalProcessedItems { get; }
-		public int TotalScheduledItems { get; }
-		public int CurrentScheduledItems { get; }
-
-		public BackgroundProcessorEventArgs(int currentScheduledItems, int totalProcessedItems, int totalScheduledItems)
-		{
-			CurrentScheduledItems = currentScheduledItems;
-			TotalProcessedItems = totalProcessedItems;
-			TotalScheduledItems = totalScheduledItems;
-		}
+		public int TotalProcessedItems { get; } = totalProcessedItems;
+		public int TotalScheduledItems { get; } = totalScheduledItems;
+		public int CurrentScheduledItems { get; } = currentScheduledItems;
+		public T Item { get; } = item;
 	}
 }