+++ ArcdpsLogManager/Uploads/DpsReportUploader.cs
@@ -22,7 +22,7 @@ namespace GW2Scratch.ArcdpsLogManager.Uploads
 			"Imperva. Supports HTTP ONLY. Fairly reliable, use as last resort.");
 
 		private static readonly DpsReportDomain DomainB = new DpsReportDomain("https://b.dps.report",
-			"Stackpath. Supports HTTPS. Alternative service domain that supports HTTPS.");
+			"Sucuri. Supports HTTPS. Alternative service domain that supports HTTPS.");
 
 		public static IReadOnlyList<DpsReportDomain> AvailableDomains { get; } = new[]
 		{